pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyspark-etl-pipeline"
version = "0.1.0"
description = "Production-grade PySpark medallion ETL pipeline for e-commerce events"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Martin", email = "martin@infin-ia.com" }]
dependencies = [
    "pyspark>=3.5,<5",
    "delta-spark>=3.2",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "faker>=24.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "chispa>=0.10",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
etl-pipeline = "pipeline.cli:main"
etl-generate = "pipeline.generator:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
ignore_missing_imports = true
packages = ["pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
